Refactor Visualization class and clean up imports
Removed unused imports and streamlined the Visualization class for better readability and maintainability. This refactor enhances the overall structure and prepares the code for future enhancements.
